requirements.txt
flask
paho-mqtt
Adafruit-PCA9685
opencv-python
numpy
Flask-MQTT
opencv-python